Adds error handling for db connections
This is mostly because when using docker-compose and the
official postgres docker image it takes a while for the db
container to start initially and the app would fail to connect and
just give up. This way we retry and handle it gracefully.

It's worth noting that the db containers only takes a while to start up
during the first run. Subsequent runs will be faster.